requirements.txt
Flask==3.0.0
Flask_SQLAlchemy==3.1.1
Flask_Migrate==4.0.5
Flask_JWT_Extended==4.6.0
python-dotenv==1.0.1
PyMySQL==1.1.0
Werkzeug==3.0.1
bcrypt==4.1.2
pytest==7.4.2
pytest-flask==1.2.0